% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_transmission.R
\name{bootstrap_transmission}
\alias{bootstrap_transmission}
\title{Bootstrap Transmission Rates}
\usage{
bootstrap_transmission(
  results,
  N_samples,
  uncorrelate = T,
  burnin = 0.1,
  plot_ci = T
)
}
\arguments{
\item{results}{MCMC output as generated by run_mcmc.}

\item{N_samples}{The number of bootstrap samples to draw.}

\item{uncorrelate}{If using correlated results, should we first uncorrelate them? (Defualts to TRUE.)}

\item{burnin}{We discard the first (burnin * 100) percent of the results, if running uncorrelation.}

\item{plot_ci}{Should we plot the bootstrap confidence intervals of the number of transmissions per person? (Defaults to TRUE.)}
}
\value{
A N_samples x (N + 1) matrix, where N is the number of hosts in the outbreak, and the additional column reports the number of index cases. Each row is one bootstrap sample of the mean number of transmissions per person (or mean number of index cases).
}
\description{
This function extracts bootstraps samples of the mean number of transmissions per person.
}
