% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bounce_rate}
\alias{bounce_rate}
\title{bounce_rate}
\usage{
bounce_rate(sessions, decimal_places = 2)
}
\arguments{
\item{sessions}{a list of sessions, generated with \code{\link{sessionise}}}

\item{decimal_places}{the number of decimal places to round the output to - set to 2 by default.}
}
\value{
a single numeric value, representing the percentage of sessions that are bounces.
}
\description{
calculates the "bounce rate" within a set of sessions - the proportion of sessions
consisting only of a single event.
}
\seealso{
\code{\link{session_events}} for generaliseable event-level calculations.
}

