% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decipher.R
\name{decipher}
\alias{decipher}
\title{Decipher MCMC Output}
\usage{
decipher(results)
}
\arguments{
\item{results}{MCMC output as generated by run_mcmc.}
}
\value{
A data frame.
}
\description{
This function shows the mapping between case numbers in MCMC output and case names.
}
