% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{SNP Distance Matrix}
\usage{
distance(results)
}
\arguments{
\item{results}{MCMC output as generated by run_mcmc.}
}
\value{
A matrix.
}
\description{
This function returns a SNP distance matrix based on the FASTA input.
}
