% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run MCMC Algorithm}
\usage{
run_mcmc(
  N_iters = 1e+05,
  filters = NULL,
  mins = NULL,
  maxs = NULL,
  vars = NULL,
  seed = T
)
}
\arguments{
\item{N_iters}{Number of MCMC iterations.}

\item{filters}{Filters for input data. NULL = defaults.}

\item{mins}{Minimum values of model parameters. NULL = defaults.}

\item{maxs}{Maximum values of model parameters. NULL = defaults.}

\item{vars}{Variances of MCMC moves for model paramaters. NULL = defaults.}

\item{seed}{Whether or not to set the seed, for replicable results. Defaults to TRUE.}
}
\value{
A list, where each entry is the current state of the Markov chain (multiples of 100 steps).
}
\description{
This function executes the MCMC algorithm for outbreak reconstruction.
}
