% Generated by roxygen2 (4.0.2): do not edit by hand
\name{session_dataset}
\alias{session_dataset}
\title{Example event dataset}
\format{a data.frame of 61,667 rows consisting of:
\describe{
  \item{UUID}{Hashed and salted unique identifiers representing 10,000 unique clients.}
  \item{timestamp}{timestamps, pre-converted using \code{\link{to_seconds}} to be immediately usable by \code{\link{sessioniser}}}
}}
\description{
an example dataset of events, drawn from the Wikimedia web properties, for experimenting with
when performing session reconstruction and analysis
}

