% Generated by roxygen2 (4.0.2): do not edit by hand
\name{session_events}
\alias{session_events}
\title{session_events}
\usage{
session_events(sessions)
}
\arguments{
\item{sessions}{a list of sessions generated via \code{\link{sessionise}}}
}
\description{
count the number of events in a session (or set of sessions)
}
\details{
\code{session_events} counts the number of events in a session, or in multiple
sessions, based on a provided "sessions" list (which can be generated via \code{\link{sessionise}})).
}
\examples{
\dontrun{
#With a sessionised dataset (see ?sessionise for an example)
event_counts <- session_events(sessions)
}
}
\seealso{
\code{\link{sessionise}} for generating sessions, \code{\link{session_length}}
for session length, and \code{\link{bounce_rate}} for the bounce rate represented by a set
of sessions.
}

