% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sessionise}
\alias{sessionise}
\title{sessionise}
\usage{
sessionise(timestamps, threshold = 3600L)
}
\arguments{
\item{timestamps}{a list of vectors of second values. These can be extracted by converting
POSIXlt or POSIXct timestamps into numeric elements through a call to \code{\link{as.numeric}},
or with the assistance of \code{\link{to_seconds}}.}

\item{threshold}{the threshold, in seconds, for splitting out a new session. Set to
3600 (one hour) by default.}
}
\value{
a list of vectors, each one representing a single session.
The returned list can be conveniently
passed into \code{\link{session_length}} or \code{\link{session_events}} for further analysis.
}
\description{
split a series of timestamps (or a series of series) associated with UUIDs, into sessions
}
\details{
\code{sessionise} splits timestamps associated with user events into "sessions",
enabling the simple calculation of various metrics such as session length or the number of events
within a session.
}
\seealso{
\code{\link{session_length}} for calculating session length, and
\code{\link{session_events}} for calculating the number of pages in each session.
}

