% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{sessionreconstruct}
\alias{sessionreconstruct}
\alias{sessionreconstruct-package}
\title{foo}
\description{
sessionreconstruct provides functions to aid in reconstructing and analysing user sessions.
Although primarily designed for web sessions (see the 'session reconstruction methodology' vignette), its session
approach is plausibly applicable to other domains, and the
analytical functions also provided in the package should be generalisable.
}
\author{
Oliver Keyes <okeyes@wikimedia.org>
}

