% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncorrelate.R
\name{uncorrelate}
\alias{uncorrelate}
\title{Extract Uncorrelated Posterior Samples}
\usage{
uncorrelate(results, burnin = 0.1)
}
\arguments{
\item{results}{MCMC output as generated by run_mcmc.}

\item{burnin}{We discard the first (burnin * 100) percent of the results}
}
\value{
An uncorrelated Markov chain, of the same format as the "results" input.
}
\description{
This function extracts uncorrelated posterior samples of an outbreak.
}
