% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize}
\alias{visualize}
\title{Visualize MCMC Output}
\usage{
visualize(results, burnin = 0.1)
}
\arguments{
\item{results}{MCMC output as generated by run_mcmc.}

\item{burnin}{We discard the first (burnin * 100) percent of the results}
}
\value{
Three plots.
}
\description{
This function plots the log-likelihood, transmission network graph, and transmission probability chart.
}
