% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{reconstructr}
\alias{reconstructr}
\alias{reconstructr-package}
\title{functions for session reconstruction and analysis}
\description{
sessionreconstruct provides functions to aid in reconstructing and analysing user sessions.
Although primarily designed for web sessions (see the introductory vignette), its session
approach is plausibly applicable to other domains.
}
\author{
Oliver Keyes <okeyes@wikimedia.org>
}

