% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/session_dataset.R
\name{session_dataset}
\alias{session_dataset}
\title{Example event dataset}
\format{a data.frame of 63,524 rows consisting of:
\describe{
  \item{UUID}{Hashed and salted unique identifiers representing 10,000 unique clients.}
  \item{timestamp}{timestamps, convertable via \code{\link{to_seconds}} (see the examples there)}
}}
\description{
an example dataset of events, drawn from the Wikimedia web properties, for experimenting with
when performing session reconstruction and analysis
}

