% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{session_events}
\alias{session_events}
\title{count the number of events in a session (or set of sessions)}
\usage{
session_events(sessions)
}
\arguments{
\item{sessions}{a list of sessions generated via \code{\link{reconstruct_sessions}}}
}
\description{
\code{session_events} counts the number of events in a session, or in multiple
sessions, based on a provided "sessions" list (which can be generated via \code{\link{reconstruct_sessions}})).
}
\examples{
\dontrun{
#With a sessionised dataset (see ?reconstruct_sessions for an example)
event_counts <- session_events(sessions)
}
}
\seealso{
\code{\link{reconstruct_sessions}} for generating sessions, \code{\link{session_length}}
for session length, and \code{\link{bounce_rate}} for the bounce rate represented by a set
of sessions.
}

